user can choose an address and checkout

+++ app/controllers/profile/addresses_controller.rb
@@ -12,8 +12,33 @@ class Profile::AddressesController < ApplicationController
 
   def destroy
     address = Address.find(params[:id])
-    address.destroy
-    redirect_to profile_path
+    address.orders.each do |order|
+      if (order.status == 'shipped') || (order.status == 'packaged')
+        flash[:danger] = "Cannot delete an address associated with a completed order."
+        redirect_to profile_path
+      elsif (order.status == 'cancelled') || (order.status == 'pending')
+        address.orders.clear
+        address.destroy
+        redirect_to profile_path
+      end
+    end
+  end
+
+  def edit
+    @address = Address.find(params[:id])
+  end
+
+  def update
+    address = Address.find(params[:id])
+    address.orders.each do |order|
+      if (order.status == 'shipped') || (order.status == 'packaged')
+        flash[:danger] = "Cannot update an address associated with a completed order."
+        redirect_to profile_path
+      elsif (order.status == 'cancelled') || (order.status == 'pending')
+        address.update(address_params)
+        redirect_to profile_path
+      end
+    end
   end
 
   private
